Share common browser logic via "internal external" interfaces
