+++ navigation/navigation-runtime/src/webMain/kotlin/androidx/navigation/BrowserHistory.kt
@@ -17,9 +17,109 @@
 package androidx.navigation
 
 import androidx.core.bundle.Bundle
+import kotlinx.coroutines.DelicateCoroutinesApi
+import kotlinx.coroutines.channels.awaitClose
+import kotlinx.coroutines.coroutineScope
+import kotlinx.coroutines.flow.Flow
+import kotlinx.coroutines.flow.callbackFlow
+import kotlinx.coroutines.launch
+
+/**
+ * Bind the browser navigation state to the given navigation controller.
+ *
+ * @param navController an instance of NavController handling the navigation logic
+ */
+internal suspend fun BrowserWindow.bindToNavigation(navController: NavController) {
+    coroutineScope {
+        val localWindow = this@bindToNavigation
+        val appAddress = with(localWindow.location) { origin + pathname }.removeSuffix("/")
+        var initState = true
+        var updateState = true
+
+        launch {
+            localWindow.popStateEvents().collect { event ->
+                val state = event.state.toString()
+
+                val restoredRoutes = state.lines()
+                val currentBackStack = navController.currentBackStack.value
+                val currentRoutes = currentBackStack.filter { it.destination !is NavGraph }
+                    .mapNotNull { it.getRouteWithArgs() }
+
+                var commonTail = -1
+                restoredRoutes.forEachIndexed { index, restoredRoute ->
+                    if (index >= currentRoutes.size) {
+                        return@forEachIndexed
+                    }
+                    if (restoredRoute == currentRoutes[index]) {
+                        commonTail = index
+                    }
+                }
+
+                //don't handle next navigation calls
+                updateState = false
+
+                if (commonTail == -1) {
+                    //clear full stack
+                    currentRoutes.firstOrNull()?.let { root ->
+                        navController.popBackStack(root, true)
+                    }
+                } else {
+                    currentRoutes[commonTail].let { lastCommon ->
+                        navController.popBackStack(lastCommon, false)
+                    }
+                }
+
+                //restore stack
+                if (commonTail < restoredRoutes.size - 1) {
+                    val newRoutes = restoredRoutes.subList(commonTail + 1, restoredRoutes.size)
+                    newRoutes.forEach { route -> navController.navigate(route) }
+                }
+            }
+        }
+
+        launch {
+            navController.currentBackStack.collect { stack ->
+                if (stack.isEmpty()) return@collect
+
+                val routes = stack.filter { it.destination !is NavGraph }
+                    .map { it.getRouteWithArgs() ?: return@collect }
+
+                val newUri = "$appAddress/${routes.last()}"
+                val state = routes.joinToString("\n")
+
+
+                if (updateState) {
+                    if (initState) {
+                        localWindow.history.replaceState(state, "", newUri)
+                        initState = false
+                    } else {
+                        localWindow.history.pushState(state, "", newUri)
+                    }
+                }
+                updateState = true
+            }
+        }
+    }
+}
+
+@OptIn(DelicateCoroutinesApi::class)
+@Suppress("UNCHECKED_CAST_TO_EXTERNAL_INTERFACE")
+private fun BrowserWindow.popStateEvents(): Flow<BrowserPopStateEvent> = callbackFlow {
+    val localWindow = this@popStateEvents
+    val callback: (BrowserEvent) -> Unit = { event: BrowserEvent ->
+        if (!isClosedForSend) {
+            (event as? BrowserPopStateEvent)?.let { trySend(it) }
+        }
+    }
+
+    localWindow.addEventListener("popstate", callback)
+    awaitClose {
+        localWindow.removeEventListener("popstate", callback)
+    }
+}
 
 private val argPlaceholder = Regex("""\{*.\}""")
-internal fun NavBackStackEntry.getRouteWithArgs(): String? {
+private fun NavBackStackEntry.getRouteWithArgs(): String? {
     val entry = this
     val route = entry.destination.route ?: return null
     if (!route.contains(argPlaceholder)) return route
@@ -34,4 +134,29 @@ internal fun NavBackStackEntry.getRouteWithArgs(): String? {
             acc.replace("{$argumentName}", value)
         }
     return routeWithFilledArgs.takeIf { !it.contains(argPlaceholder) }
+}
+
+internal external interface BrowserLocation {
+    val origin: String
+    val pathname: String
+}
+
+internal external interface BrowserHistory {
+    fun pushState(data: String?, title: String, url: String?)
+    fun replaceState(data: String?, title: String, url: String?)
+}
+
+internal external interface BrowserEvent
+internal external interface BrowserPopStateEvent : BrowserEvent {
+    val state: String?
+}
+
+internal external interface BrowserEventTarget {
+    fun addEventListener(type: String, callback: ((BrowserEvent) -> Unit)?)
+    fun removeEventListener(type: String, callback: ((BrowserEvent) -> Unit)?)
+}
+
+internal external interface BrowserWindow : BrowserEventTarget {
+    val location: BrowserLocation
+    val history: BrowserHistory
 }
